arreglos, +1000 feels de que se acabara boku no pico temporada 2...

+++ Gran Torismo API/Gran Torismo API/Models/Establecimientos.cs
@@ -40,7 +40,7 @@ namespace Gran_Torismo_API.Models
         public string descripcion { get; set; }
 
         [BsonElement("precio")]
-        public double precio { get; set; }
+        public int precio { get; set; }
 
         [BsonElement("fotos")]
         public BsonArray fotos { get; set; }
